Extracted the anonymous worker subclasses to their own files to make the code more readable.
